# BKSEC training 2026/low-effort-sns/low-effort sns 2/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A social networking platform designed for developers and automation enthusiasts. Connect, share, and interact through API-driven interactions.",
        "title": "API-SNS",
        "contact": {},
        "version": "1.33.7"
    },
    "host": "loweffsns2.bksec.vn",
    "basePath": "/",
    "paths": {
...
        "/secret": {
            "get": {
                "description": "Fetches the secret associated with the currently logged-in user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Secret"
                ],
                "summary": "View own secret",
                "responses": {
                    "200": {
                        "description": "User's secret",
                        "schema": {
                            "$ref": "#/definitions/model.Secret"
                        }
                    },
                    "404": {
                        "description": "User does not have a secret",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/secret/{id}": {
            "get": {
                "description": "Fetches the secret associated with a specific user based on the provided ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Secret"
                ],
                "summary": "View another user's secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's secret",
                        "schema": {
                            "$ref": "#/definitions/model.Secret"
                        }
                    },
                    "404": {
                        "description": "User does not have a secret",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Secret": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "description": "JWT Key for Authorization. Example: Authorization: Bearer XXXX",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}